requirements.txt
cffi==1.15.1
ffprobe==0.5
future==0.18.2
hachoir==3.1.3
numpy==1.23.1
Pillow==9.2.0
pycparser==2.21
